requirements.txt

flask
flask-cors
psycopg2
pyjwt
pyyaml
